refactor openmc conversion code to move to folder
