+++ pythonAPI/omegah2csg/convert2openmc/__init__.py
@@ -1,1 +1,29 @@
+from .convert2openmc import (
+    is_horizontal,
+    is_vertical,
+    get_slope,
+    get_vertical_intersection,
+    get_line_equation,
+    is_cone,
+    create_z_cone,
+    create_z_plane,
+    create_z_cylinder,
+    create_openmc_surface,
+    create_openmc_geometry,
+    convert2openmc,
+)
 
+__all__ = [
+    "is_horizontal",
+    "is_vertical",
+    "get_slope",
+    "get_vertical_intersection",
+    "get_line_equation",
+    "is_cone",
+    "create_z_cone",
+    "create_z_plane",
+    "create_z_cylinder",
+    "create_openmc_surface",
+    "create_openmc_geometry",
+    "convert2openmc",
+]
